chore: create use_veto_vote test scenario
